+++ api-types/src/lib.rs
@@ -24,8 +24,8 @@ pub use restore::{
 };
 use serde::Deserialize;
 pub use vm_config::cpus_config::{
-    CoreScheduling, CpuAffinity, CpuFeatures, CpuTopology, CpuTopologyParseError,
-    ParseCoreSchedulingError,
+    CoreScheduling, CpuAffinity, CpuFeatures, CpuTopology, CpuTopologyParseError, CpusConfig,
+    CpusConfigParseError, ParseCoreSchedulingError,
 };
 pub use vm_coredump::VmCoredumpData;
 pub use vm_remove_device::VmRemoveDeviceData;
