+++ vmm/src/config.rs
@@ -12,10 +12,9 @@ use std::str::FromStr;
 use std::sync::LazyLock;
 
 use api_types::{
-    CoreScheduling, CpuAffinity, CpuFeatures, MemoryRestoreMode, RestoredNetConfig,
-    RestoredVfioConfig, VmMemoryZoneUpdateData,
+    CpusConfig, CpusConfigParseError, MemoryRestoreMode, RestoredNetConfig, RestoredVfioConfig,
+    VmMemoryZoneUpdateData,
 };
-use arch::CpuProfile;
 use block::ImageType;
 use clap::ArgMatches;
 use log::warn;
@@ -92,10 +91,7 @@ pub enum Error {
     ParseVsockCidMissing,
     /// Error parsing CPU options
     #[error("Error parsing --cpus")]
-    ParseCpus(#[source] OptionParserError),
-    /// Invalid CPU features
-    #[error("Invalid feature in --cpus features list: {0}")]
-    InvalidCpuFeatures(String),
+    ParseCpus(#[from] CpusConfigParseError),
     /// Error parsing memory options
     #[error("Error parsing --memory")]
     ParseMemory(#[source] OptionParserError),
@@ -668,104 +664,6 @@ impl FromStr for HotplugMethod {
             "virtio-mem" => Ok(HotplugMethod::VirtioMem),
             _ => Err(ParseHotplugMethodError::InvalidValue(s.to_owned())),
         }
-    }
-}
-
-impl CpusConfig {
-    pub fn parse(cpus: &str) -> Result<Self> {
-        let mut parser = OptionParser::new();
-        parser
-            .add("boot")
-            .add("max")
-            .add("topology")
-            .add("kvm_hyperv")
-            .add("max_phys_bits")
-            .add("affinity")
-            .add("features")
-            .add("nested")
-            .add("core_scheduling")
-            .add("profile");
-        parser.parse(cpus).map_err(Error::ParseCpus)?;
-
-        let boot_vcpus: u32 = parser
-            .convert("boot")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or(DEFAULT_VCPUS);
-        let max_vcpus: u32 = parser
-            .convert("max")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or(boot_vcpus);
-        let topology = parser.convert("topology").map_err(Error::ParseCpus)?;
-        let kvm_hyperv = parser
-            .convert::<Toggle>("kvm_hyperv")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or(Toggle(false))
-            .0;
-        let max_phys_bits = parser
-            .convert::<u8>("max_phys_bits")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or(DEFAULT_MAX_PHYS_BITS);
-        let affinity = parser
-            .convert::<TupleList<u32, Vec<usize>>>("affinity")
-            .map_err(Error::ParseCpus)?
-            .map(|v| {
-                v.0.iter()
-                    .map(|Tuple(e1, e2)| CpuAffinity {
-                        vcpu: *e1,
-                        host_cpus: e2.clone().into_boxed_slice(),
-                    })
-                    .collect()
-            });
-
-        let profile = parser
-            .convert::<CpuProfile>("profile")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or_default();
-
-        let features_list = parser
-            .convert::<StringList>("features")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or_default();
-
-        #[allow(unused_mut)]
-        let mut features = CpuFeatures::default();
-        {
-            #[cfg(target_arch = "x86_64")]
-            for feature in features_list.0 {
-                match feature.as_str() {
-                    "amx" => features.amx = true,
-                    _ => return Err(Error::InvalidCpuFeatures(feature)),
-                }
-            }
-
-            #[cfg(not(target_arch = "x86_64"))]
-            if let Some(feature) = features_list.0.into_iter().next() {
-                return Err(Error::InvalidCpuFeatures(feature));
-            }
-        }
-
-        let nested = parser
-            .convert::<Toggle>("nested")
-            .map_err(Error::ParseCpus)?
-            .is_none_or(|toggle| toggle.0);
-
-        let core_scheduling = parser
-            .convert("core_scheduling")
-            .map_err(Error::ParseCpus)?
-            .unwrap_or(CoreScheduling::Vm);
-
-        Ok(CpusConfig {
-            boot_vcpus,
-            max_vcpus,
-            topology,
-            kvm_hyperv,
-            max_phys_bits,
-            affinity,
-            features,
-            nested,
-            core_scheduling,
-            profile,
-        })
     }
 }
 
@@ -3786,110 +3684,6 @@ mod unit_tests {
     use net_util::MacAddr;
 
     use super::*;
-
-    #[test]
-    fn test_cpu_parsing() -> Result<()> {
-        assert_eq!(CpusConfig::parse("")?, CpusConfig::default());
-
-        assert_eq!(
-            CpusConfig::parse("boot=1")?,
-            CpusConfig {
-                boot_vcpus: 1,
-                max_vcpus: 1,
-                ..Default::default()
-            }
-        );
-        assert_eq!(
-            CpusConfig::parse("boot=1,max=2")?,
-            CpusConfig {
-                boot_vcpus: 1,
-                max_vcpus: 2,
-                ..Default::default()
-            }
-        );
-        assert_eq!(
-            CpusConfig::parse("boot=8,topology=2:2:1:2")?,
-            CpusConfig {
-                boot_vcpus: 8,
-                max_vcpus: 8,
-                topology: Some(CpuTopology {
-                    threads_per_core: 2,
-                    cores_per_die: 2,
-                    dies_per_package: 1,
-                    packages: 2
-                }),
-                ..Default::default()
-            }
-        );
-
-        CpusConfig::parse("boot=8,topology=2:2:1").unwrap_err();
-        CpusConfig::parse("boot=8,topology=2:2:1:x").unwrap_err();
-        assert_eq!(
-            CpusConfig::parse("boot=1,kvm_hyperv=on")?,
-            CpusConfig {
-                boot_vcpus: 1,
-                max_vcpus: 1,
-                kvm_hyperv: true,
-                ..Default::default()
-            }
-        );
-        assert_eq!(
-            CpusConfig::parse("boot=2,affinity=[0@[0,2],1@[1,3]]")?,
-            CpusConfig {
-                boot_vcpus: 2,
-                max_vcpus: 2,
-                affinity: Some(Box::new([
-                    CpuAffinity {
-                        vcpu: 0,
-                        host_cpus: Box::new([0, 2]),
-                    },
-                    CpuAffinity {
-                        vcpu: 1,
-                        host_cpus: Box::new([1, 3]),
-                    }
-                ])),
-                ..Default::default()
-            },
-        );
-
-        // Test core_scheduling parsing
-        assert_eq!(
-            CpusConfig::parse("boot=1,core_scheduling=vm")?,
-            CpusConfig {
-                boot_vcpus: 1,
-                max_vcpus: 1,
-                core_scheduling: CoreScheduling::Vm,
-                ..Default::default()
-            }
-        );
-        assert_eq!(
-            CpusConfig::parse("boot=1,core_scheduling=vcpu")?,
-            CpusConfig {
-                boot_vcpus: 1,
-                max_vcpus: 1,
-                core_scheduling: CoreScheduling::Vcpu,
-                ..Default::default()
-            }
-        );
-        assert_eq!(
-            CpusConfig::parse("boot=1,core_scheduling=off")?,
-            CpusConfig {
-                boot_vcpus: 1,
-                max_vcpus: 1,
-                core_scheduling: CoreScheduling::Off,
-                ..Default::default()
-            }
-        );
-        // Default (no core_scheduling specified) should be Vm
-        assert_eq!(
-            CpusConfig::parse("boot=1")?.core_scheduling,
-            CoreScheduling::Vm
-        );
-        // Invalid value should error
-        CpusConfig::parse("boot=1,core_scheduling=invalid").unwrap_err();
-
-        Ok(())
-    }
 
     #[test]
     fn test_mem_zone_parsing() -> Result<()> {
